add sugar selection

+++ smart_coffee_machine/Pot.h
@@ -6,7 +6,7 @@ class Potentiometer{
     public:
         Potentiometer(int pin);
 
-        float getValue();
+        int getValue();
 
     private:
         int pin;
